Vulnerability

Documentation for Vulnerability.

Vulnerability.raytracing

Vulnerability.shotlines

Vulnerability.target

Vulnerability.target

Vulnerability.raytracing — Method

raytracing(fragments;args)

Returns a list of fragments and a list of rays. The rays represent the fragments trajectories and can be used to intersect targets. Optional argument is the v_threshold. Which is a threshold velocity used to discard fragments for which the velocity is lower. The default value is 0.0.

Vulnerability.shotlines — Method

shotlines(target,rays)

Returns the number of hit and a list of rays that interserct with the target.

Vulnerability.target — Method

target(target,xlength,ylength,zlength)

Creates a cuboid target of dimensions (xlength, ylength and zlength) at the target position for shotline intersection computation

Vulnerability.target — Method

target(size)

Creates a cuboid target of dimensions specified by the vector size